requirements.txt
Pillow
PySide6
pyautogui
pygetwindow
